LAApatch: moved code to a new file and tweaked some of the interface
